requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.0.6
    # via kombu
asgiref==3.4.1
    # via django
beautifulsoup4==4.10.0
    # via -r requirements.in
billiard==3.6.4.0
    # via celery
celery==5.1.2
    # via -r requirements.in
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
django==3.2.8
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.12.4
    # via -r requirements.in
kombu==5.1.0
    # via celery
prompt-toolkit==3.0.20
    # via click-repl
psycopg2==2.9.1
    # via -r requirements.in
pytz==2021.3
    # via
    #   celery
    #   django
redis==3.5.3
    # via -r requirements.in
six==1.16.0
    # via click-repl
soupsieve==2.2.1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
xmltodict==0.12.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
